feat(addie): auto-reconcile Slack ↔ WorkOS account links daily
Extracts the bulk email-matching logic from the admin POST endpoint into
a shared autoLinkUnmappedSlackUsers() in slack/sync.ts, then registers a
daily background job that runs it. The job fires 2 minutes after startup
to fix existing unmapped users (like Matt Egol) and every 24 hours to
catch future stragglers.
