+++ .changeset/slack-auto-link-job.md
@@ -1,0 +1,4 @@
+---
+---
+
+Add daily background job to reconcile unmapped Slack users to website accounts by email, running 2 minutes after startup and every 24 hours thereafter.
